requirements.txt
psycopg2-binary
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
joblib
fastapi
uvicorn
streamlit
requests
beautifulsoup4
selenium
python-telegram-bot
prophet
imbalanced-learn
python-dotenv
plotly==6.5.0
